Fixed some path issues when running on Windows
The APP_DIR constant was changed to use os.homedir(), as that function
is inherently crossplatform.

Also made a change in createFile. Initially it was separating paths
using a separator that broke the file creation process on Windows.
Instead of using the forward slash token, it now splits using the token
located in path.sep.